Fixing ManufacturerDataMask bug
Fixing bug initializing the ManufacturerDataMask

Bytes of the DataMask are initialised with 0b00000001 instead of
0b11111111, hence only the least significant bits of the filter and the
result set are being compared. Effectively the FilterMask is just an
even/odd filter for the selected bytes.

Also added the iBeacon type identifier (0x02) as well as length byte
(0x21) to the filter mask. I understand this is done in the
onScanResult method @ScannerScanCallback, however, as type byte and
length are fixed this will reduce